# Save expensive interface type checks (related JDK-8180450 and https://github.com/netty/netty/issues/12708)