JDK-8180450 has forced Netty http encoding/decoding to order type checks in order to save a huge scalability hit.
And still this is not enough, perf-wise: vertx, as a Netty users, still consume Netty types and cannot ignore such ordering, or would have a similar performance hit.

Moreover, there are cases (outlined on https://github.com/TechEmpower/FrameworkBenchmarks/pull/7942) where the existing JDK mechanism for interface type checks waste CPU resources, but the "constrained" vertx http types can save it to happen both performing exact class type checks and/or instanceof vs concrete and known types.