@franz1981 2023-02-23
I've run the existing JMH benchmark `HttpServerHandlerBenchmark` but:
- sadly it's limited to just one side of the http handling ie `ertxHttpRequestDecoder` but **not** `VertxHttpResponseDecoder`
- metrics are enabled (!) and it is badly hitting a SUPER slow path type check on `DefaultHttpRequest` on https://github.com/eclipse-vertx/vert.x/blob/5ea7bc2fe2cf17ab78b57c5ac68386821bb841b5/src/main/java/io/vertx/core/http/impl/Http1xConnectionBase.java#L195-L226

I'll modify the benchmark in order to make evident the benefits