@franz1981 2023-02-28
Before this PR:
```
Benchmark                                      Mode  Cnt    Score   Error   Units
HttpServerHandlerBenchmark.vertx              thrpt   20  614.545 ± 7.614  ops/ms
HttpServerHandlerBenchmark.vertxOpt           thrpt   20  777.142 ± 7.255  ops/ms
HttpServerHandlerBenchmark.vertxOptMetricsOn  thrpt   20  759.129 ± 7.972  ops/ms
```
These are the results post PR:

```
Benchmark                                      Mode  Cnt    Score    Error   Units
HttpServerHandlerBenchmark.vertx              thrpt   20  640.783 ±  4.690  ops/ms
HttpServerHandlerBenchmark.vertxOpt           thrpt   20  834.770 ± 3.142  ops/ms
HttpServerHandlerBenchmark.vertxOptMetricsOn  thrpt   20  832.764 ±  4.015  ops/ms
````
While Netty (now optimized in order to save expensive type checks as well):
```
Benchmark                                      Mode  Cnt    Score    Error   Units
HttpServerHandlerBenchmark.netty              thrpt   20  917.029 ±  3.416  ops/ms
```

In short:
- this PR improve http performance ~7-10% 
- even with such improvement, at our best, we're ~10% slower then Netty